Refactoring heredoc functions

+++ redirs/heredoc.c
@@ -6,53 +6,14 @@
 /*   By: anunes-o <anunes-o@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/11 14:34:03 by anunes-o          #+#    #+#             */
-/*   Updated: 2026/03/31 16:19:03 by anunes-o         ###   ########.fr       */
+/*   Updated: 2026/04/04 16:45:33 by anunes-o         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-/* vai comparar se a situação contem ou não o heredoc e também se ele esta 
-ou não dentro de pipes, se certificando também a execução do heredoc na 
-sequência certa, também atribui os valores na struct de t_redir
-     cat << EOF
-	(fd)    (delimitador)
-*/
-int	apply_heredocs(t_ast *node)
-{
-	t_redir	*tmp;
-	int		fd;
-
-	if (!node)
-		return (0);
-	if (node->type == NODE_PIPE)
-	{
-		if (apply_heredocs(node->left) < 0)
-			return (-1);
-		if (apply_heredocs(node->right) < 0)
-			return (-1);
-		return (0);
-	}
-	else
-	{
-		tmp = node->redirs;
-		while (tmp)
-		{
-			if (tmp->type == T_HEREDOC)
-			{
-				fd = heredoc(tmp->file);
-				if (fd < 0)
-					return (-1);
-				tmp->heredoc_fd = fd;
-			}
-			tmp = tmp->next;
-		}
-	}
-	return (0);
-}
-
 /* vai apenas gerar um nome único para cada arquivo temporário do heredoc*/
-char	*generate_heredoc_name(void)
+static char	*generate_heredoc_name(void)
 {
 	static int	index;
 	char		*value;
@@ -65,47 +26,61 @@ char	*generate_heredoc_name(void)
 	return (name);
 }
 
-/* 
+/* vai popular o arquivo temporário linha por linha, até o delimiter ser 
+encontrado, pra depois heredoc poder reabrir esse arquivo em modo leitura
 */
-int	heredoc(char *delimiter)
+static int	write_line(int fd, char *line, char *filename)
 {
-	int		fd;
-	char	*line;
-	char	*filename;
-	ssize_t	ret;
+	if (write(fd, line, ft_strlen(line)) == -1 || write(fd, "\n", 1) == -1)
+	{
+		perror("write");
+		return (close_fd(fd, line, filename));
+	}
+	free(line);
+	return (0);
+}
 
-	line = NULL;
-	filename = generate_heredoc_name();
-	fd = open(filename, O_CREAT | O_WRONLY | O_TRUNC, 0644);
-	if (fd < 0)
-		return (-1);
+static	int	heredoc_loop(int fd, char *delimiter, char *filename)
+{
+	char	*line;
+
 	while (1)
 	{
+		line = readline("> ");
 		if (get_signal() == 130)
 		{
-			if (line)
-				free(line);
-			close(fd);
-			unlink(filename);
 			reset_signal();
-			return (-1);
+			return (close_fd(fd, line, filename));
 		}
-		line = readline("> ");
-		if (line == NULL)
+		if (!line)
 			break ;
 		if (ft_strcmp(line, delimiter) == 0)
 		{
 			free(line);
 			break ;
 		}
-		ret = write(fd, line, ft_strlen(line));
-		ret = write(fd, "\n", 1);
-		if (line)
-			free(line);
+		if (write_line(fd, line, delimiter) < 0)
+			return (-1);
 	}
-	close(fd);
+	return (0);
+}
+
+/* 
+*/
+int	heredoc(char *delimiter)
+{
+	int		fd;
+	char	*filename;
+
+	filename = generate_heredoc_name();
+	fd = open(filename, O_CREAT | O_WRONLY | O_TRUNC, 0644);
+	if (fd < 0)
+		return (-1);
+	if (heredoc_loop(fd, delimiter, filename) < 0)
+		return (-1);
 	fd = open(filename, O_RDONLY);
 	unlink(filename);
 	free(filename);
 	return (fd);
 }
+
